requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
fastapi==0.66.0
h11==0.12.0
idna==3.2
peewee==3.14.4
pydantic==1.8.2
PyMySQL==1.0.2
python-dotenv==0.18.0
requests==2.26.0
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
